Add `Transaction` class to manage account transactions
A new `Transaction` class was introduced in the `BankingApp.Accounts` namespace. This class includes properties for `Amount`, `Date`, `From`, and `TO` to represent transaction details. A constructor was added to initialize these properties.

The `ToString` method was overridden to provide a formatted string representation of the transaction. A `SendMail` method was defined but is currently unimplemented (marked as TODO).

Unused properties `TransactionID` and `TransactionMessage` were commented out. Necessary `using` directives were added to the file.